+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.1.1
 * Android: Fix BluetoothDevice null-safety compile error under Kotlin 2.x
+* Fix `BleCharacteristic` equality and `hashCode` to compare `properties` by value
 
 ## 2.1.0
 * Add optional `queueId` parameter to all APIs